Move TestCaseMixin to DBTestCase
